GET request completed

+++ CRUD API using Function Based View/api/app/serializer.py
@@ -3,5 +3,4 @@ from rest_framework import serializers
 class TeacherSerializer(serializers.Serializer):
     full_name = serializers.CharField(max_length=100)
     address = serializers.CharField(max_length=100)
-    salary = serializers.BigIntegerField()
-    
+    salary = serializers.IntegerField()
